refactor(Core UI): :recycle: Rename file Slide to Fade
