+++ src/pages/Addresses/views/FlowPairs.tsx
@@ -1,7 +1,7 @@
 import { useParams } from 'react-router-dom';
 
 import { AvailableProtocols } from '@API/REST.enum';
-import TransitionPage from '@core/components/TransitionPages/Slide';
+import TransitionPage from '@core/components/TransitionPages/Fade';
 
 import ConnectionsByAddress from '../components/ConnectionsByAddress';
 import RequestsByAddress from '../components/RequestsByAddress';
