+++ src/pages/shared/Loading/index.tsx
@@ -4,7 +4,7 @@ import { Bullseye, Title } from '@patternfly/react-core';
 import { CogIcon } from '@patternfly/react-icons';
 
 import { getTestsIds } from '@config/testIds.config';
-import TransitionPage from '@core/components/TransitionPages/Slide';
+import TransitionPage from '@core/components/TransitionPages/Fade';
 
 import { Labels } from './Loading.enum';
